requirements.txt
groq
fastapi==0.104.1
pydantic==2.4.2
python-dotenv==1.0.0
uvicorn==0.24.0
